[Animation Worklet] Do not allow undefined/NaN local time
Restrict localTime type so that NaN is considered invalid which brings us
in line with web-animation.

Our implementation was previously incorrect as it would accept NaN and then
convert it into an unexpected base::TimeDelta which does not make much sense.

This patch restricts accepted input by switching EffectProxy.localTime from
'unrestricted double?' -> 'double?' and brings us in line with
ComputedEffectTiming.localTime as defined in Web-Animation. [1,2]

As a side effect the attribute no longer accepts 'undefined'. Per webidl
'double?' is a nullable type which accepts a valid double, or null. So
anywhere we are setting local time to 'undefined' we should use 'null'
instead.

[1] https://drafts.csswg.org/web-animations-1/#dom-computedeffecttiming-localtime
[2] https://cs.chromium.org/chromium/src/third_party/blink/renderer/core/animation/computed_effect_timing.idl?l=10

Change-Id: I9a72e4023608980c8a64743bc6737202e4f70538