Use testdriver Action API in WPT pointerevent touch tests - Part 2
In order to run the web-platform-tests automatically, we will use
testdriver Action API in all the wpt to simulate inputs. Here we are
changing some pointerevent touch input tests which are doing scrolling.

This is the second CL that changes part of the pointerevent touch input
tests.

Bug: 606367
Change-Id: I67cc6c401a22de59af298133021f380baa7d4b19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1473971
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#645386}