Let OBJECT frames persist through reattachment when possible.
If we need to reattach any OBJECT when entering printing, we cannot wait
for the resource to load again. This was especially bad for LayoutNG,
which switches the document back to legacy layout for printing, so that
everything gets reattached, so that no OBJECT would print.

To fix this, do the same as in HTMLFrameElementBase::AttachLayoutTree()
for HTMLPluginElement, i.e. re-use ContentFrame()->View() when
reattaching, with some necessary modifications: If we end up in
LazyReattachIfNeeded(), we need to disable re-use of the old frame
view. Not sure why, exactly, but some tests would fail otherwise.

Bug: 913929
Change-Id: I3b11b0550a775c8df1c0449c2438569aa8219091