Portals: Implement PortalHost.postMessage
Implements a version of postMessage for PortalHost that can can only
send strings. It implements messaging in the opposite direction to
https://crrev.com/c/1423462 (from the portal to the embedding page).
Support for more general message types will be added in a follow up CL.

This CL introduces 2 new mojo interfaces:

blink.mojom.PortalClient: It is a counterpart to blink.mojom.Portal,
and allows messages to be sent from content::Portal in the browser
process to its corresponding HTMLPortalElement in the renderer process.

blink.mojom.PortalHost: Used by the blink::PortalHost object (in the
renderer process) to communicate with the content::Portal (in the
browser process) that owns the WebContents inside which the PortalHost
is exposed.

Both the above interfaces together allow blink::PortalHost to send
messages to the blink::HTMLPortalElement (which owns the frame inside
which the PortalHost is exposed).

Bug: 914120
Change-Id: If7dbaf6e2817fcc94dc38a1df25afc25d8b75eb3